Rework meshopt_Allocator to store callbacks in function static
Using templated class static allowed us to avoid function calls even in
unoptimized builds, but results in conflicts with Clang's C++20 module
implementation when meshoptimizer.h is included in global module
fragment.

Instead, we now use a function static to store this state and require a
function call to get access to it. The function static should, similarly
to the previous code, be instantiated in translation units that use it
and deduplicated by linker.