Add MESHOPTIMIZER_ALLOC_EXPORT configuration to share allocation storage
Before this change (and irrespectively of the previous change to
callback storage), allocator callbacks were deduplicated by the linker
within an individual module, but when meshoptimizer was compiled as a
shared library, the library code and the calling code would get
separate configuration.

This meant that meshopt_setAllocator would only affect the allocation
behavior of the library internals, but would not affect the temporary
allocations done by meshopt_IndexAdapter. To fix that, we need to make
sure the storage function can be implemented in allocator.cpp and
exported.