demo: Implement scaled error for simplification
Instead of using the same error threshold for each mesh, we scale each
threshold using the relative size of the mesh to the entire scene. The
code for this assumes no instancing or same-scale instancing, ideally
we'd have to compute the worst scale factor based on all instances.

On multi-mesh scenes this makes the thresholding significantly more
correct, both for fixed ratio and auto LOD configurations.