simplifier: Filter out zero-area triangles
When a vertex is collapsed onto a seam pair, if it was connected to both
vertices of the pair, the collapse will create a zero area triangle
which is not topologically degenerate. These triangles aren't removed by
remapIndexBuffer so we need to filter them out.

While it would be best to do this filtering during every pass so that
the simplifier accurately tracks the remaining triangles, doing this
breaks loop metadata in a way that is difficult to repair, which can
hinder further collapses around the same loop due to loop checks
failing. So for now we just do this at the end of the simplification.