gltfpack: Scale simplification error with mesh extents
Instead of using the same target error for each mesh, we now use a
factor derived from the ratio of mesh extent to the largest mesh extent
to ensure that meshes of different sizes get roughly the same visual
degradation.

When meshes are instanced, we take the largest possible scale as a
factor; smaller instances will retain more detail.

While we could use the ratio of mesh bounds to scene bounds instead of a
ratio to max mesh bounds, there's cases where the input is a collection
of objects that are intended to be moved around in space, and the scene
bounds is not necessarily representative. With this the simplification
behavior is roughly similar to quantization behavior (modulo scale
treatment).