simplifier: Compute precise component nexterror while pruning
Instead of computing nexterror based on component data, do it based on
the remaining indices. This is more or less free, and ensures that while
nexterror can be stale after unrelated collapses (that may effectively
remove a component edge by edge), calling pruneComponents restores
nexterror to be up to date.