simplifier: Fix early out during final cleanup pass
In some meshes, by the time we get to the final cleanup pass, the low
error components have all been removed through edge collapses but
component_nexterror was not updated since the reported error was lower.
There might still be higher error components that are eligible for
pruning, but we run the risk of stopping after one iteration of pruning
if the removed component had sufficiently low error.

We now make sure that early out doesn't trigger on the first iteration;
pruneComponents will update the error precisely now, and if we are still
stuck we will leave the loop. The early out is useful as a way to ensure
the loop always terminates regardless of possible issues around fp
comparisons and NaNs.