demo: Display mixed edges in orange in debug overlay
In some cases loop metadata may be returned for edges that can't
currently be collapsed; for example, currently some vertices are
misclassified as border and have a border-seam loop that is effectively
locked, but the edges used to be rendered as a seam edge, suggesting it
may be collapsible.

For now we assume any edge with mismatching non-locked kinds can't be
collapsed and render them using a separate color.