overdraw: Compute mesh centroid using vertex data
Instead of iterating through indices when computing mesh centroid,
iterate through vertices. This results in averaging a mostly
deduplicated set of positions, which produces a slightly different
result but the overall overdraw results are within margin of error on a
large mesh set.

However, this results in much faster computation of the centroid itself
as we need to process fewer points and do so with linear memory access,
so this makes optimizeOverdraw ~15% faster overall on typical meshes.