Use older Ubuntu version for release builds
GHA supports Ubuntu 22 and 24 LTS; we currently use 24 (latest) to build
release binaries, which picks up a dependency on newer glibc. This is
optional, so with this we switch to older Ubuntu. Similarly, switch to
explicit version for macOS so that we don't accidentally regress
runnability.

For regular builds, macOS version for ARM builds was hardcoded and
already one version behind; to avoid this version getting constantly
stale, use -latest here.