simplifier: Account for triangle filtering when finalizing grid size
Binary search for the target grid size is performed using an approximate
calculation for triangle count which overestimates the result as it does
not filter out duplicate triangles. This often creates cases where a
slightly larger grid would still fit in the triangle budget.

This could be solved by using post-filtered triangle count during the
binary search but that's very expensive. Most often, the excess is small
enough that we can only bump the grid size by 1, so instead we perform a
single probe pass to check if increasing the grid size is possible.

To do this we use the same code that we use for triangle filtering, but
use vertex ids instead of vertex cells; it is still correct because
there's a 1-1 relationship between these, and while the destination
buffer becomes non-sensical briefly, it's only used as temporary storage
during hashing.

To avoid reallocating the hash table and keep the flow simple, we only
do this if the hash table is sufficiently large to perform the probing.