simplifier: Expand the triangle filter hash table a bit
Instead of allocating enough storage for min_triangles, we allocate
enough for target triangle count; this reduces the chance of skipping
the probing pass - before this it would happen ~10% of the time and now
it happens ~1%. The extra cost of the adjustment is insignificant.

Also fix tracing of the probe grid size.