simplifier: Expand triangle filter hash table dynamically
Instead of trying to predict the triangle filter hash table size ahead
of time, we can size it to accommodate both current and slightly larger
grid for probing. This requires us to run countTriangles to get the
worst case bound, but we need to do this anyhow to do filtering safely.
This makes sure that we can filter a probed grid even on small grid
sizes, where significant gains from probing are the most likely.