simplifier: Implement support for vertex locking for simplifySloppy
Since each vertex is mapped to a cell, if a vertex is mapped to a unique
cell, that vertex will stay in place. Additionally, any triangle with
two locked vertices will not be collapsed as the third vertex will move
to a non-locked cell. This makes it easy to provide vertex locking
support similar to the support in the main simplifier: each locked
vertex is mapped to a special vertex id which carries the vertex index
as unique payload.

Unlike meshopt_simplify, we do not have a wedge map here; because of
this, all vertices with the same position must be consistently locked to
prevent holes in the mesh.