demo: Add wireframeOverlay as an alternative to wireframe
This makes it much easier to see the tessellation on complex meshes
without making the mesh see-through (which is still valuable to see
internal classification/structure in some cases).

Also remove the double call to simplify when debug overlay is enabled;
we can retrieve the original index buffer by stripping the debug bits,
which doubles the performance on complex meshes.