simplifier: Use point quadrics for regularization
In addition to face and edge quadrics, we now introduce point quadrics,
that encode distance to the vertex position. We used to use these for
point simplification but that was replaced with quadric-less metric for
simplicity instead, but they can be used to improve mesh simplification.

For planar regions, regular triangle quadrics evaluate to 0 for any
point on the plane. This leads to chaotic triangulation as the collapse
order is effectively random, and can produce skinny triangles which is
suboptimal for rasterization performance. Additionally, quadrics like
this can't be minimized.

Adding point quadrics to vertex quadrics with a very small factor
improves both issues significantly. The factor is chosen to minimally
affect the overall simplification error, but might be adjusted in the
future.