Update README.md
Deemphasize meshopt_simplifySloppy: we still mention this in "basic
simplification" section, but remove the code example and spend less text
on it. Elsewhere we focus on the main simplification algorithm and its
various entry points. In the future, more options will be implemented to
make sloppy algorithm even less relevant.