demo: Fix mergeVertices for .obj files with ~large extents
mergeVertices clamps tolerance to 1e-16 and divides by it, and proceeds
to truncate the result to 32-bits; as such, using low tolerances is
unsafe, especially on models with reasonably large extents, as it
increases the likelihood of a spurious hash collision. Use 1e-4 default
instead of 0 to work around this.

Also skip unindexed meshes just in case glTF input contains them.