vertexfilter: Implement initial version of color filter
Colors have inherent dependency between different components: changing
luminance ends up changing all three components and the resulting
entropy needs to be encoded separately per channel. Additionally, when
using sub-byte bit counts, it can be difficult to arrange the bits to
compress well, remain compatible with normalized integer decoding, and
preserve endpoints.

This change introduces a color filter that performs a lossy encoding by
converting the color to YCoCg. YCoCg is easier to decode compared to YUV
or other formats. Since YCoCg requires more range for Co & Cg, we
truncate that range and as such the filter is lossy.

To be able to encode different elements at different bit counts, which
is a capability all other filters have, we encode the bit count using
the most significant 1 bit in the alpha channel; as a result, this
encoding is limited to 7 (or 15) bits in the alpha channel. Often alpha
is constant and vertex alpha is somewhat less important so this is not a
significant loss in practice.