vertexfilter: Decode YCoCg using integer arithmetics
While the existing solution works and can be converted to SIMD as is,
the result is a little slower than optimal - ~10% slower than octahedral
encoding despite being less sophisticated.

To fix this, we need to adjust the decoding to be able to operate at the
same scale, instead of using two scaling factors. That alone helps a
little bit, but crucially allows us to move the entire decoding to
integer arithmetics.

This still requires the decoder to clamp the decoded value to the
expected range. For now, instead we encode the values in such a way that
the decoder never over- or under-flows by using YCoCb-R encoding with
truncation.