vertexfilter: Implement SSE2 decoding for color filter
The code structure here is quite similar to octahedral decoding; instead
of decoding the inputs to floats and computing the entire result with
floating point math, we perform RGB conversion using integers.

For scaling factor, when using 16-bit colors, rcp precision is not quite
sufficient so we resort to a full division. 8-bit version can use
mul+rcp.