vertexfilter: Implement Wasm SIMD version of decodeColorFilter
We follow SSE implementation as is, with the exception of integer
conversion: while we only need an unsigned float to integer rounded
conversion here, it's still faster to use the floating point rounding
trick instead of adding 0.5 and using convert.

Unlike SSE, we use full division in 8-bit decoding as well as there's no
easy access to reciprocal estimate.

Note: as of LLVM 19, this hits a compiler bug with computing the scale
factor, where convert_i32x4_s is canonicalized to convert_i32x4_u since
'as' is known to have zero bit cleared. We might need to work around
this in the future.