vertexfilter: Work around LLVM misoptimizing convert_i32x4_s
LLVM 19 started using known bits to detect if the argument to
convert_i32x4_s is non-negative and canonicalizing it to
convert_i32x4_u. This can significantly reduce performance on x64, where
a performant u32->f32 cast is not available until AVX512, and engines
like v8 do not support that code path. On color filters this results in
20-25% performance regression.

Work around this by using volatile masks. For color filter, this
unfortunately requires an extra instruction and thus the performance is
still not peak but it's better than doing nothing, and this workaround
can be removed when LLVM fixes it.

On the positive side, the old workaround for another codegen issue is no
longer necessary.