simplifier: Enhance edge error with triangle quadrics
Using edge quadrics for border and seam edges counterintuitively results
in degraded quality on the collapses around the loop. This is due to
normalized error: accumulating a quadric with a large weight shifts the
error distribution away from previously accumulated quadrics. This is
fine for edges that are in otherwise planar regions, but results in
unnaturally low error for collapsing vertices along the loop.

For now, fix this by re-applying triangle quadrics with the same weight,
but don't incorporate the weight during addition. This effectively
results in an approximation of distance-to-edge as opposed to
distance-to-edge-along-triangle-surface, which more accurately captures
the error.

Note that the two quadrics are weighted slightly differently; in the
future this might need to be adjusted, but it might also be nice to have
some freedom for anisotropic behavior.