simplifier: Experiment with isotropic error on seam edges
Instead of using two planes with potentially different weights, we can
use a true distance-to-line formula, which can be encoded into a
quadric.

The code was derived by taking the canonical distance to a line
length(cross(p1 - p0, p1 - XYZ)), and expanding the squared version to
compute coefficients for X^2/XY/etc.

For ease of weighting, the direction is normalized prior to encoding;
the code works just as well without a normalize as long as weight is
divided by squared length, but that does not improve precision and
normalize is easier to follow.

It's unclear if this is better or worse than anisotropic version.