simplifier: Use anisotropic quadrics applied twice on seams
While using distance to line seems to work fine for the most part, it's
not obviously better than anisotropic quadrics; sometimes visual
apperance is a little better on anisotropic quadrics and they give us an
ability to control the degree to which plane vs edge matter.

So for now, revert to anisotropic quadrics but apply them properly on
seams: while seams are bidirectional, the quadrics we end up computing
depend on the adjoining triangle which is different for each side. To
maintain existing weighting, this requires using 0.5 as the default
weight - which is subject to future tuning.