simplifier: Classify half-seams as seams instead of borders
In complex meshes it's possible to encounter a somewhat unusual type of
vertex: one that only has a single copy (wedge), and that has one
incoming and one outgoing open edge, but both edges go to a vertex with
the same position.

Previously, we classified vertices like this as border, because border
classification ignored a possibility that both endpoints to to the same
vertex. This created cases where a border vertex is adjacent to a seam
vertex, but neither can be collapsed onto each other because of collapse
restrictions.

Doing the right thing here is somewhat complicated. Locking vertices
like this to avoid misleading classification regresses cases where a
half-seam vertex like this is connected to a locked vertex; these are
the cases that in theory may lead to issues on discontinuities but in
practice we don't seem to observe them. Classifying this as a new type
does not solve that either, as deciding that this vertex can move leads
to similar issues in theory.

For now we re-classify vertices like this as seams. The seam handling
logic is already safe in case a seam vertex only has a single copy,
modulo an assertion that this change adjusts. If this leads to problems
down the line, the next best thing is to lock these, at least until we
have new complex collapse rules in place.