simplifier: Expose an option for quadric solve
This change just adds an option as well as basic scaffolding to
recompute positions and attributes without a full solve. For now the
option is "exposed" similarly to debug visualization: it's available via
a hidden option bit.

To correctly recompute attributes, we need to keep the attribute remap
around so that, similarly to sparse remap, we can go back to the
original attribute index from a compacted one.