simplifier: Implement basic position quadric solve
For now we are using Cramer's rule to directly invert the quadric. Most
quadrics should be well behaved after regularization, however it's still
not the most precise or efficient method but it serves as a proof of
concept.

We update all positions, even though some may not be used; this can be
fixed after the implementation is more complete.