simplifier: Simplify LDL^T computation a little bit
The code before used the literal LDL decomposition algorithm; there's a
little bit of redundancy there as it computes a product of L factors and
multiplies it by D, but L was already divided by D; we can use the
numerator directly in cases like this which saves us a few multiplies.

The computation is now structured to be able to take advantage of fma if
available.

Also move di checks to the end of the function; the intermediate results
may produce inf/nan which is unimportant.