simplifier: Only solve manifold/complex vertices for now
Any vertex with an attribute discontinuity can result in significant
visual issues when solved due to UV reprojection. We currently optimize
the position with just a single wedge attribute, but even if we
accumulate all the wedges, the resulting positions on a UV seam
generally move the position laterally. Reprojecting the UV on a
resulting position then results in UV that is outside of the original UV
chart, causing significant issues in the appearance.

For borders, solving them requires a much larger edge penalty so that
the vertex stays closer to the border line; borders are also not as
valuable to optimize.