simplifier: Only use solved position if it doesn't introduce flips
Similarly to edge collapse, updating the position may result in an
adjacent triangle flipping. To avoid this we use the same adjacency
structure and predicate that we already have built to validate
positions.

Also fix a missed check for vertex_update[] in attribute update loop
which resulted in wasted work.