simplifier: Use more granular classification rules for quadric solve
We still only allow position update for manifold or complex vertices;
however, we now update attributes on any non-locked vertices, which
improves normal quality on seams.

To move borders, we'll need to use stronger edge quadrics so that the
geometry is preserved better. For seams, we either need to use a very
strong edge quadric to ensure UVs don't need to extrapolate, or find a
different way to do attribute optimization that preserves UVs better.

Locked vertices are currently left without an update; this is important
for positional update for similar reasons to seams, and for attributes
we can't update externally locked vertices (to avoid introducing seams)
and we don't currently distinguish external locking vs internal locking.