simplifier: Compute and update volume gradients for solve
By itself, solving quadric metric with attributes tends to reduce volume
compared to the use of the position metric. However, position metric
alone doesn't produce optimally placed vertices which results in worse
attributes in certain cases.

To fix this, we need to constraint the minimization with an extra volume
gradient; this change computes it in a separate array, only when
attributes are used together with solve. The gradient is not used during
solve yet.