simplifier: Integrate volume gradient into quadric solve
We can now solve the quadric with an additional linear constraint. The
solve process is the same - we use LDL decomposition and solve
individual smaller systems via substitution.

If the constraint does not exist or is poorly conditioned, the extra D
element of the resulting decomposition will be small. Instead of failing
the entire solve, we disable the constraint by using 0 for the relevant
L/D elements, which ends up computing the same result as our original
solve.

This approximately doubles the cost of quadricSolve, but it does produce
better positions when attributes are enabled, so for now this is
probably the best solution.