Change cifuzz GHA action to run only on pushes for longer
Fuzzing build alone takes 2 minutes because of Google OSS container
infrastructure. This makes this build delay PR checks for 3 minutes
which doesn't make a lot of sense given that this hasn't found issues
previously.

Instead, let's run this on pushes to master only, and run the actual
fuzzing for longer. This should bring the build time up to about 5
minutes but it's not blocking and should not fail often.