gltfpack: Implement support for interleaved vertices
When a new command line argument -vi is used, gltfpack now produces
interleaved vertex streams instead of deinterleaving per attribute.

Interleaving is not compatible with vertex filters, so they are disabled
in this mode. This reduces compression efficiency when options like -cc
or -vpf together with -c are used. Filters can still be used on
animation data.

When the mesh is using morph targets, each morph target is interleaved
separately. This prevents vertices from having very large strides on
meshes with many targets, and results in more efficient access patterns
if a small subset of morph targets is used at runtime.