simplifier: Use stricter bounds on quadric decomposition
Instead of permitting 1e-7 threshold for elements of D when solving the
quadric, change this to 1e-6. 1e-7 seems to be too close to the useful
floating point precision boundary (it's slightly below 1.0f ULP!), and
can sometimes misclassify a solve that we should not go forward with.

With default regularization weight of 1e-7, this has a byproduct of
refusing to solve planar regions - however, they don't really *need* to
be solved anyway, and if planar regions got a higher regularization
weight the solve would proceed without issues.