simplifier: Add another regularization term during solve
While neighborhood bounds are effective at removing the invalid position
updates, they do not guarantee that the position we've solved is good.
Additionally, if the update is good but lies on the neighborhood
boundary, we may see discontinuous surface if some vertices update and
some others don't. Thus ideally we'd have another complementary solution
addressing solve issues.

When full regularization is enabled, we empirically don't see any
instability or wrong position updates; however, the default
regularization weight (1e-7) is insufficient and significantly
increasing it will affect the resulting errors. This change thus adds a
regularization quadric just during solve, and just for the target vertex
- this is a little different from regularization during initialization
as the extra quadrics injected there accumulate.

The extra weight is selected to be large enough to fully stabilize the
update during testing, but small enough to not affect the position
visibly when well formed constraints are present. For simplicity, we add
this even if full regularization is enabled; it's not really affecting
the results in that case.