Mark partitionClusters and spatialClusterPoints as stable
Both functions work well within their intended function at this point,
so they can be stabilized. partitionClusters may see further
implementation improvements in the future that wouldn't affect the
interface.

It's possible that we'd want to extend the interface for
partitionClusters in the future - for example, to specify min/max range
for partition sizes, or some tuning parameters - but we already have an
existing paradigm for this with buildMeshletsFlex which we can use if
necessary here.