Update README.md
Update the list of experimental functions to exclude recently stabilized
APIs and add simplifyWithUpdate and SimplifyRegularize which are new.

Also with the list of experimental APIs being much shorter now, the
"likely to become stable" moniker is somewhat unnecessary and going
forward there shouldn't really be a strong difference between
"experimental" and "experimental but will be stabilized soon", as
experimental functions may change regardless.