simplifier: Remove zero area triangles iteratively
During collapses, we sometimes create zero area triangles that are not
topologically degenerate. This can happen when collapsing manifold
vertices onto seam pairs (such as when folding a thin cylindrical shape
onto its own seam spine) and for some other types of collapses onto
locked vertices. Additionally, while this is not enabled right now, this
happens very often when vertices are tagged as Complex as these
collapses may cross wedge boundaries.

What happens in this case right now is that while we do clean up these
triangles at the end of simplification, they stick around until the very
end. This may not be a problem if there's only a few edges, but in
meshes with many edges like this that may throw off the final triangle
target, producing under-detailed meshes. In cases where simplification
ratio is significant, this can penalize the final mesh too much: on
buddha.obj, for 100x simplification we are left with 10% (2k/20k) of
degenerate triangles at the end. This issue becomes critical if Complex
collapses are introduced, as delaying filtering until the end produces
poor quality results.

The reason why we previously deferred this until the very end is that it
breaks the loop metadata, which can't be easily restored in
remapEdgeLoops; a separate filtering pass would be necessary to identify
broken loop edges and properly terminate them. However, a loop edge that
leads to an unconnected vertex is functionally identical to a ~0u tag -
neither will match against an edge. A loop break like this does *not*
prevent collapsing the seam vertex further along the loop away from the
break, and does *not* prevent from fully collapsing the loop segment
into a point.

Thus at this point it's better to proceed with removing these triangles
early. We could add loop metadata cleanup in the future if it becomes
necessary; for now, with that we don't need a separate filter function
as we can adjust the degenerate condition in remapIndexBuffer instead.