simplifier: Evaluate worst case attribute error for complex collapses
When we collapse Complex vertices, we collapse the entire wedge complex
onto a single target vertex. This requires adjusting the attribute
error: for seams, we add up the errors as measured towards the
respective target vertex of the pair, whereas here the entire complex is
evaluated against a single target vertex instead.

Unlike typical wedge loops, here we already have accounted for the base
collapse target so we can use a regular for loop.