simplifier: Improve opposite edge handling for complex vertices
Previously, manifold->complex edges were not considered to have an
opposite edge; this would lead to redundant duplicate evaluation as we'd
visit both edges and add one edge as forward, and the opposite edge as
reverse, which are identical. This could also lead to overflowing the
collapse buffer and early discard of some collapses.

As for complex->complex edges, while they usually also occur twice in
the attribute topology, the two edges have different wedge endpoints for
faceted regions; because complex collapses collapse onto the same wedge,
leaving both variants in the queue makes sure that we rank both and
select the optimal target automatically.