simplifier: Streamline edge loop collapse checks
Instead of using a separate branch for border/seam->locked collapses, we
now check each edge with one border/seam vertex for whether it lies on
the corresponding loop. The check is omitted if the other vertex is a
manifold, as manifold->border/seam collapses are allowed from outside of
the loop.

These are exactly equivalent: for seam-seam/border-border edges, the
loop metadata is symmetric and as such the other check must pass as
well, and for seam/border->locked we simply replicate the previous
logic. However, this code is simpler and supports seam/border->complex
collapses out of the box (if they were to be enabled). This also matches
the way we validate edges in fillEdgeQuadrics.