js: Expose permissive mode as a setting
This is very experimental so it does not show up in type definitions or
documentation yet.

Also remove redundant Uint8Array creation for vertex_lock arguments; it
dates back to an earlier experimental version where this was a JS array,
but it's been a Uint8Array for a while now.