demo: Implement permissive fallback for hierarchical simplification
If a given cluster fails to simplify below the threshold, we repeat the
simplification attempt using permissive mode; before the entire process begins,
we mark vertices that lie on UV seams as protected.

Unlike sloppy simplification, this results in reasonable appearance and despite
some current limitations of permissive mode, the fallback strategy allows us
to generate much deeper DAGs for faceted meshes and other complex topology.

The protection code is a little strange because the remap[] table we use doesn't
remap vertices to other vertices, so we have to build an "unremap" table to get
a single canonical vertex to compare all others against.