indexgenerator: Implement experimental meshopt_generatePositionRemap
In several algorithms, it's useful to have a position remap that maps
vertices to the "canonical" (first) vertex with the same position. This
is somewhat similar to shadow indexing, and indeed such a remap can be
used to reindex buffers for shadow rendering, but this is also useful
for simplification to compute protection or lock bits, and for cluster
partitioning.

It's almost possible to use meshopt_generateVertexRemap for this, but in
addition to requiring that the input count is divisible by 3, this
function doesn't handle negative zeroes; this is fine for remaps wrt
efficiency but can lead to rare issues where vertices with the same
position aren't protected for simplification which could lead to
geometry or attribute artifacts.

A new dedicated function solves both cases cleanly. It uses the same
hasher we use for generateVertexRemapCustom, just without the callback.