gltfpack: Introduce an option for permissive simplification
This is meant as an alternative to aggressive mode, which is using
sloppy simplification as a fallback. Sloppy simplification doesn't
preserve attributes and doesn't support attribute weighting, and has
issues with volume and shape conservation.

In contrast, permissive mode can simplify across discontinuities and can
do it in a directed fashion. This change exposes permissive mode in its
most basic form, without the use of protection bits; these will be added
in a separate commit.