gltfpack: Protect UV seams and sharp creases in permissive mode
This change uses vertex_lock to set protection bits on vertices on UV
and sharp normal discontinuities. The UV discontinuities may get folded
otherwise as we do not currently pass them to the simplifier; while this
is technically optional for normals, collapsing complex edges on a
normal crease results in an attribute fold with a high error, which can
reduce quality or block these collapses depending on the error limit.