gltfpack: Implement support for building Basis Universal via CMake
As Basis Universal develops further, a single-file source build gets
more and more difficult to maintain, especially given that the set of
source files changes with version (which we don't pin in an obvious way)
- but also due to various incompatibilities with single source.

This change introduces a new way to build gltfpack with Basis Universal
compression via CMake. Since our existing build setup relies on the
source option, for now this support is implemented side-by-side; the
older configuration will be disabled.