gltfpack: Use CMake Basis builds and split macOS builds by arch
Instead of using make BASISU=..., which is a development configuration,
use CMake to build gltfpack with all optional dependencies in build.yml.
Note that gltfpack-coverage build still relies on Makefile for now.

For release configuration, we switch to using Basis CMake builds instead
of single-source file builds. This, as well as a future libwebp support,
requires us to split the macOS builds by arch due to SIMD configuration
- for Basis, -D SSE is required for Intel macOS builds but not for ARM,
and libwebp which will be added in the future has similar configuration
issues. Intel builds will be deprecated at some point anyway, so for now
we split the builds.