gltfpack: Implement PNG/JPEG decode for WebP encode via Basis
When gltfpack is built with Basis Universal support, we have easy access
to jpegd/pv_png decoders that are used for KTX2 texture encoding. This
is very helpful for WebP support, because libwebp imagedec relies on
libpng/libjpeg which need to be downloaded and built separately; this
allows us to bypass that dependency automatically.