gltfpack: Use higher quality for WebP normal maps
We now remap quality parameter a little bit based on the texture type;
for now, normal maps use a different ramp, which changes the defaults
(tq 8 maps to quality 90 for normal maps and quality 84 for color maps,
instead of 80 for both previously). These can be adjusted in the future
along with some other parameters perhaps.