gltfpack: Use external basis_universal build in Makefile
This removes the last non-optional dependency on basislib.cpp, allowing
us to remove it. Instead we rely on externally building
libbasisu_encoder.a and simply link to it.

This also is helpful because we can use optimized build for basisu and a
debug build of gltfpack during development if necessary.