gltfpack: Use JPEG-like quality control for WebP encoding
The default curve produces small and low-quality files for much of the
earlier levels, and the useful range seems to be in the higher end. By
comparison, enabling JPEG-like quality mode flattens out the curve
somewhat and seems to be more comparable to other formats.

This has a side-effect of increasing the size of files produced by
default, with corresponding quality increases...