simplifier: Only compute protected flags for primary vertices
For permissive processing of protected flags, the results for all wedges
of the same vertex are the same; by nature of this processing we always
have at least 2, and sometimes 3 or more, wedges, so we're just doing
more work without any reason.

This can be structured to process each wedge sequentially, but for now
it's easier to keep the existing code structure and only process the
primary vertex instead. This reduces classifyVertices cost by ~2-3x
depending on the topology.