js: Fix possibly stale heap access in computeSphereBounds
We need to allocate all memory via sbrk and then fill it all using the
new heap view. Otherwise there's a small possibility that the call to
sbrk the radii array will reallocate the heap and the heap.set call will
fail as the view is stale and the write will be out of bounds.