simplifier: Introduce meshopt_SimplifyVertex_* flags for vertex locks
Instead of documenting numeric constants like 0x01/0x02, expose the
relevant bit masks as enum values. Note that the enum entries are just
called "Vertex_", with no mention of locks/flags, as there's some
ambiguity in the parameter name which may or may not be resolved in the
future.

The enum flags can't really change - 0x01 must be the lock bit because
that retains ABI/API compatibility with existing code using boolean
values, and 0x02 will have to be the protect bit going forward for
forward API/ABI compatibility - but using named constants makes the code
easier to understand.